requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
trl>=0.7.4
accelerate>=0.25.0
vllm>=0.1.0
rich>=13.7.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# API and model dependencies
openai>=1.0.0
huggingface-hub>=0.19.0
comet>=1.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
aiofiles>=23.2.0

# Optional: for bfloat16 training
triton>=2.1.0 
